Rename ResultsMap.js to MapContainer.js
